added comments
