+++ test/index.html
@@ -29,7 +29,7 @@
     <script src="../js/settings.js"></script>
     <script src="../js/atb.js"></script>
     <script src="../js/tab.js"></script>
-    <script src="../js/tabs.js"></script>
+    <script src="../js/tabManager.js"></script>
     <script src="../js/trackerLists.js"></script>
     <script src="../js/list_importers/importers.js"></script>
     <script src="../js/list_importers/disconnect.js"></script>
